+++ fjs/basen/proof.f.ts
