+++ fjs/basen/todo/various-basen-encodings.md
